requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.2
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
duplicity==0.8.21
exceptiongroup==1.2.0
fastapi==0.104.1
fasteners==0.14.1
filelock==3.13.1
frozenlist==1.4.0
future==0.18.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.25.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonpatch==1.33
jsonpointer==2.4
keyring==23.5.0
langchain==0.0.340
langsmith==0.0.66
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
loguru==0.7.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.20.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
openai==1.3.5
packaging==23.2
pandas==2.1.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pinecone-client==2.2.4
platformdirs==3.11.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.6.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2023.10.3
reportlab==3.6.8
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
ssh-import-id==5.11
starlette==0.27.0
systemd-python==234
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.24.0.post1
virtualenv==20.24.6
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.9.3
zipp==1.0.0
PyJWT==2.8.0
aiohttp-retry==2.8.3
twilio==9.0.0